+++ lib/iex/api.rb
@@ -19,3 +19,4 @@ end
 
 require_relative 'api/quote'
 require_relative 'api/company'
+require_relative 'api/price'