+++ lib/iex/resources.rb
@@ -1,3 +1,4 @@
 require_relative 'resources/resource'
 require_relative 'resources/quote'
 require_relative 'resources/company'
+require_relative 'resources/price'